+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Allow multiple regulators to be added to a profession
+- Display nations below organisations in search results
 
 ### Changed